Optimize Code and Finis architecture

+++ prisma/migrations/20241019115446_name_back/migration.sql
@@ -3,7 +3,7 @@ CREATE TABLE `Player` (
     `id` INTEGER NOT NULL AUTO_INCREMENT,
     `name` VARCHAR(191) NOT NULL,
     `country` VARCHAR(191) NOT NULL,
-    `money` DECIMAL(65, 30) NOT NULL DEFAULT 0.0,
+    `money` INTEGER NOT NULL DEFAULT 0,
 
     PRIMARY KEY (`id`)
 ) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci;
